refactor: separate Rust file
