+++ src-tauri/src/main.rs
@@ -1,12 +1,10 @@
 // Prevents additional console window on Windows in release, DO NOT REMOVE!!
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
-use font_kit::source::SystemSource;
-use font_kit::sources::{fs::FsSource, multi::MultiSource};
+mod mojimachi;
+
 use serde::Serialize;
-use directories::UserDirs;
-use ttf_parser::name::Table;
-use ttf_parser::Tag;
+use ttf_parser::{Tag, name::Table};
 
 #[derive(Serialize)]
 struct FontInfo {
@@ -14,39 +12,13 @@ struct FontInfo {
     postscript_name: Option<String>,
 }
 
-#[cfg(target_os = "windows")]
-fn get_additional_path() -> Option<FsSource> {
-    if let Some(user_dirs) = UserDirs::new() {
-        let home_dir = user_dirs.home_dir().to_str().unwrap();
-        return Some(FsSource::in_path(home_dir.to_string() +"\\AppData\\Local\\Microsoft\\Windows\\Fonts\\"));
-    }
-
-    None
-}
-
-#[cfg(not(target_os = "windows"))]
-fn get_additional_path() -> Option<FsSource> {
-    None
-}
-
-fn get_source() -> MultiSource {
-    let other_source = get_additional_path().unwrap();
-    let source = SystemSource::new();
-    MultiSource::from_sources(
-        vec![
-            Box::new(source),
-            Box::new(other_source),
-        ]
-    )
-}
-
 #[tauri::command]
-fn get_families(keyword: Option<String>) -> Vec<String> {
-    let source = get_source();
+fn get_families(keyword: String) -> Vec<String> {
+    let source = mojimachi::get_source();
     let families = source.all_families().unwrap();
     let mut filtered_families = Vec::new();
     for family in families {
-        if let Some(keyword) = &keyword {
+        if keyword != String::from("") {
             if family.to_lowercase().contains(&keyword.to_lowercase()) {
                 filtered_families.push(family);
             }
@@ -62,8 +34,8 @@ fn get_families(keyword: Option<String>) -> Vec<String> {
 }
 
 #[tauri::command]
-fn get_ja_families(keyword: Option<String>) -> Vec<String> {
-    let source = get_source();
+fn get_ja_families(keyword: String) -> Vec<String> {
+    let source = mojimachi::get_source();
     let fonts = source.all_fonts().unwrap();
     let mut filtered_families = Vec::new();
     for font in fonts {
@@ -71,7 +43,7 @@ fn get_ja_families(keyword: Option<String>) -> Vec<String> {
         let family_name = font_object.family_name().to_string();
         let glyph = font_object.glyph_for_char('あ');
         if glyph.is_some() && glyph.unwrap() != 0 {
-            if let Some(keyword) = &keyword {
+            if keyword != String::from("") {
                 if family_name.to_lowercase().contains(&keyword.to_lowercase()) {
                     filtered_families.push(family_name);
                 }
@@ -89,7 +61,7 @@ fn get_ja_families(keyword: Option<String>) -> Vec<String> {
 
 #[tauri::command]
 fn get_fonts_info() -> Vec<FontInfo> {
-    let source = get_source();
+    let source = mojimachi::get_source();
     let all_fonts = source.all_fonts().unwrap();
     let mut fonts = Vec::new();
 
@@ -108,7 +80,7 @@ fn get_fonts_info() -> Vec<FontInfo> {
 #[tauri::command]
 fn get_fonts_head() -> Vec<Vec<Option<String>>> {
     let name_table_tag = Tag::from_bytes(b"name").as_u32();
-    let source = get_source();
+    let source = mojimachi::get_source();
     let all_fonts = source.all_fonts().unwrap();
     let mut fonts = Vec::new();
 
@@ -143,7 +115,7 @@ fn get_fonts_head() -> Vec<Vec<Option<String>>> {
 #[tauri::command]
 fn get_font_head(name: String) -> Vec<Option<String>> {
     let name_table_tag = Tag::from_bytes(b"name").as_u32();
-    let source = get_source();
+    let source = mojimachi::get_source();
     let font_handle = source.select_by_postscript_name(&name).unwrap();
     let font_object = font_handle.load().unwrap();
     let name_table_bytes = font_object.load_font_table(name_table_tag).unwrap();
@@ -171,7 +143,7 @@ fn get_font_head(name: String) -> Vec<Option<String>> {
 
 #[tauri::command]
 fn get_fonts_by_family(family: String) -> Vec<Option<String>> {
-    let source = get_source();
+    let source = mojimachi::get_source();
     let family_handle = source.select_family_by_name(&family).unwrap();
     let fonts = family_handle.fonts();
     let mut family_fonts = Vec::new();
